cogs/guildstate.py
from discord.ext import commands

class GuildState(commands.Cog):
  """Helper class managing per-guild state."""

  def __init__(self):
    self.volume = 1.0
    self.playlist = []
    self.skip_votes = set()
    self.now_playing = None

  def is_requester(self, user):
    return self.now_playing.requested_by == user