requirements.txt
aiohttp==3.5.4
apiclient==1.0.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
datadog==0.29.3
decorator==4.4.0
discord.py==1.2.3
google==2.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
oauth2client==4.1.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pymongo==3.9.0
PyNaCl==1.3.0
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.2
websockets==6.0
yarl==1.3.0
youtube-dl==2019.8.2
